Rename discrete module to analytic
